feat: replace Ollama with Google Gemini Flash for chat intelligence
Replace the local Ollama LLM (qwen3:4b) with Google Gemini 2.5 Flash API.
The local model was unreliable at producing structured JSON output for task
creation and scheduling on 4GB VRAM hardware.

- Add @google/generative-ai SDK and create gemini.ts streaming client
- Collapse two-tier model system (fast/capable) into single tier
- Remove Ollama dependencies, config, and host networking from Docker
- Remove /no_think directive from system prompts (Qwen-specific)
- Fix schedule diff showing committed blocks as "REMOVED" when they
  were simply untouched by a new schedule generation
