+++ .env.example
@@ -15,10 +15,8 @@ NEXTAUTH_URL=http://localhost:3000
 AUTH_GOOGLE_ID=
 AUTH_GOOGLE_SECRET=
 
-# ─── Ollama (Local LLM) ──────────────────────────────────────────────────────
-OLLAMA_BASE_URL=http://localhost:11434
-OLLAMA_MODEL=qwen3:4b
-# Fast model for simple queries (greetings, acks). Pull with: ollama pull qwen3:1.7b
-OLLAMA_MODEL_FAST=qwen3:1.7b
-# Context window size — lower values save VRAM (8192 balances VRAM and prompt capacity)
-OLLAMA_NUM_CTX=8192
+# ─── Google Gemini (Cloud LLM) ────────────────────────────────────────────────
+# Get your API key at: https://aistudio.google.com/apikey
+GEMINI_API_KEY=your-gemini-api-key
+# Optional: override model (default: gemini-2.5-flash)
+# GEMINI_MODEL=gemini-2.5-flash
