+++ docker-compose.prod.yml
@@ -31,12 +31,7 @@ services:
       AUTH_GOOGLE_ID: ${AUTH_GOOGLE_ID}
       AUTH_GOOGLE_SECRET: ${AUTH_GOOGLE_SECRET}
       TOKEN_ENCRYPTION_KEY: ${TOKEN_ENCRYPTION_KEY}
-      OLLAMA_BASE_URL: ${OLLAMA_BASE_URL}
-      OLLAMA_MODEL: ${OLLAMA_MODEL}
-      OLLAMA_MODEL_FAST: ${OLLAMA_MODEL_FAST}
-      OLLAMA_NUM_CTX: ${OLLAMA_NUM_CTX:-8192}
-    extra_hosts:
-      - "host.docker.internal:host-gateway"
+      GEMINI_API_KEY: ${GEMINI_API_KEY}
     depends_on:
       db:
         condition: service_healthy
