+++ package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@auth/drizzle-adapter": "^1.11.1",
+    "@google/generative-ai": "^0.24.1",
     "@tailwindcss/oxide-linux-x64-gnu": "^4.2.2",
     "clsx": "^2.1.1",
     "drizzle-orm": "^0.45.1",
