+++ src/lib/chat/planner-prompt.ts
@@ -5,8 +5,8 @@ export function buildSimpleSystemPrompt(userName: string): string {
   return `You are Runekeeper, a warm and concise weekly planning assistant.
 Respond naturally and briefly. The user's name is ${userName}.
 Do not include JSON, code blocks, or action instructions.
-Reference the "Enchanted Archivist" theme subtly (e.g., "quests" for tasks, "map" for schedule).
-/no_think`;
+Reference the "Enchanted Archivist" theme subtly (e.g., "quests" for tasks, "map" for schedule).`;
+
 }
 
 export function buildSystemPrompt(context: {
@@ -112,6 +112,6 @@ ${weekOverview}
 8. When the user says "confirm", "commit", "looks good", "yes", "do it", or clicks "Confirm plan" — immediately include confirm_plan in actions. Do NOT ask for confirmation again. Just commit and confirm it's done.
 9. Do NOT repeatedly ask "Would you like me to confirm/commit?" — if the user already confirmed, act on it.
 10. Always include a "notes" field for each task you create. If the user provides a description, use it verbatim. Otherwise, infer a brief description (1-2 sentences) from the conversation context or the task title. The description should clarify WHAT the quest involves.
+`;
 
-/no_think`;
 }
